+++ apps/web/src/app.test.tsx
@@ -31,6 +31,10 @@ const clientWith = (
 ): AppApi => ({
   logout,
   getMe: () => Promise.reject(new Error('getMe is not stubbed in this file')),
+  getThread: () => Promise.reject(new Error('getThread is not stubbed in this file')),
+  postComment: () => Promise.reject(new Error('postComment is not stubbed in this file')),
+  updateComment: () => Promise.reject(new Error('updateComment is not stubbed in this file')),
+  deleteComment: () => Promise.reject(new Error('deleteComment is not stubbed in this file')),
   login: () => Promise.reject(new Error('login is not stubbed in this file')),
   getAdminAccounts: () => Promise.reject(new Error('getAdminAccounts is not stubbed in this file')),
   setAccountRoles: () => Promise.reject(new Error('setAccountRoles is not stubbed in this file')),